Fix password resets security against DB leaks
Password reset tables are basically random passwords that allows complete access to an user account. It should not happen, but in case an attack get's access to the database, we are currently storing emails and access tokens in pure plaintext.

Tokens should be treated as passwords to prevent simple SQL dumps for allowing access to the recovery feature. The quick timeout of the tokens somewhat limits this attack, but I don't see a reason NOT to secure against it anyway. 

Paragonie talks about this also: https://paragonie.com/blog/2016/09/untangling-forget-me-knot-secure-account-recovery-made-simple#secure-password-reset-tokens